pyproject.toml
[project]
name = "paved-path"
version = "0.1.0"
description = "Library for working with files with automatic caching."
readme = "README.md"
authors = [
    { name = "ryn-cx", email = "84663588+ryn-cx@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
yaml = [
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pylint>=3.3.7",
    "pytest>=8.4.1",
    "ruff>=0.12.2",
]

[tool.ruff.lint]
select = ["ALL"] # Enable all linters.
ignore = [
    # S603 and S602 overlap where one or the other is always true. S603 is disabled
    # because S602 is more secure than S603.
    "S603",

    # This rule makes it impossible to access URLs that are dynamically created.
    "S310",

    # Docstrings should exist when they are useful, requiring them just leads to bad
    # docstrings.
    "D100",
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    #
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"] # Asserts should be allowed in tests.

[tool.pylint.'MESSAGES CONTROL']
disable = [
    # Automatically disabled by lint-roller.
    "E0237",
    "E1142",
    "E0014",
    "E1300",
    "E0013",
    "E1310",
    "E1307",
    "E2502",
    "E6005",
    "E6004",
    "E0116",
    "E0108",
    "E0241",
    "E1303",
    "E0102",
    "E0100",
    "E0605",
    "E0604",
    "E0304",
    "E0308",
    "E2510",
    "E2513",
    "E2514",
    "E2512",
    "E2515",
    "E0309",
    "E0305",
    "E0303",
    "E1206",
    "E1205",
    "E0704",
    "E1304",
    "E1302",
    "E4703",
    "E0213",
    "E0107",
    "E0115",
    "E0117",
    "E0103",
    "E0711",
    "E0643",
    "E0402",
    "E1132",
    "E0106",
    "E0101",
    "E0104",
    "E1519",
    "E1520",
    "E0114",
    "E0001",
    "E1306",
    "E1305",
    "E0112",
    "E1301",
    "E0603",
    "E0602",
    "E0302",
    "E0601",
    "E0118",
    "E1700",
    "E0105",
    "W1401",
    "W0129",
    "W0199",
    "W3201",
    "W1302",
    "W0311",
    "W1501",
    "W0211",
    "W0702",
    "W0711",
    "W1502",
    "W0718",
    "W0719",
    "W0640",
    "W0160",
    "W0102",
    "W0705",
    "W0109",
    "W0130",
    "W1641",
    "W0123",
    "W0122",
    "W0106",
    "W1309",
    "W0511",
    "W1515",
    "W1305",
    "W1310",
    "W0604",
    "W0603",
    "W0602",
    "W1404",
    "W0406",
    "W1405",
    "W1508",
    "W1113",
    "W1202",
    "W1203",
    "W1201",
    "W0150",
    "W1518",
    "W0410",
    "W1303",
    "W0131",
    "W0177",
    "W3301",
    "W2402",
    "W0133",
    "W0104",
    "W0212",
    "W0707",
    "W0622",
    "W2901",
    "W1406",
    "W0404",
    "W0127",
    "W1509",
    "W1510",
    "W0245",
    "W0706",
    "W0012",
    "W2301",
    "W0108",
    "W0107",
    "W0301",
    "W1514",
    "W0613",
    "W1304",
    "W1301",
    "W0611",
    "W0612",
    "W0120",
    "W2101",
    "W2601",
    "W0401",
    "C0202",
    "C0198",
    "C1901",
    "C0201",
    "C0501",
    "C0206",
    "C0199",
    "C0112",
    "C0415",
    "C2701",
    "C0103",
    "C0301",
    "C2201",
    "C0115",
    "C0304",
    "C0116",
    "C0114",
    "C0410",
    "C0321",
    "C2403",
    "C2401",
    "C0205",
    "C0121",
    "C0305",
    "C0303",
    "C0131",
    "C0105",
    "C0132",
    "C0412",
    "C0123",
    "C3002",
    "C2801",
    "C3001",
    "C0113",
    "C1802",
    "C0207",
    "C0208",
    "C0414",
    "C0411",
    "C0413",
    "R0133",
    "R0124",
    "R6003",
    "R1701",
    "R6002",
    "R6104",
    "R1717",
    "R0402",
    "R1728",
    "R1715",
    "R1714",
    "R1730",
    "R1731",
    "R1718",
    "R1722",
    "R1706",
    "R1732",
    "R5501",
    "R2044",
    "R1710",
    "R0123",
    "R2004",
    "R0202",
    "R1723",
    "R1724",
    "R1720",
    "R1705",
    "R6301",
    "R0203",
    "R0206",
    "R1704",
    "R1719",
    "R1703",
    "R1725",
    "R1260",
    "R0913",
    "R0916",
    "R0917",
    "R0912",
    "R0914",
    "R1702",
    "R0904",
    "R0911",
    "R0915",
    "R1707",
    "R1721",
    "R1729",
    "R1733",
    "R1734",
    "R1735",
    "R1734",
    "R1737",
    "R6201",
    "R0205",
    "R0022",
    "R1711",

    # Pylance/Pyright make these redundant because they are more accurate and faster.
    "E1101",
]
